Moved service error catch
