+++ src/main/scripts/src/app/core/http-client.service.ts
@@ -22,6 +22,8 @@ import {Injectable} from '@angular/core';
 import {Http, Headers, RequestOptions, Response, Request, RequestOptionsArgs} from '@angular/http';
 
 import {Observable} from 'rxjs/Observable';
+import 'rxjs/add/operator/catch';
+import 'rxjs/add/observable/throw';
 
 @Injectable()
 export class HttpClient {
@@ -30,36 +32,55 @@ export class HttpClient {
   }
 
   public get(url: string, requestOptions?: RequestOptionsArgs): Observable<Response> {
-    return this.http.get(url, requestOptions);
+    return this.http.get(url, requestOptions)
+      .catch(this.handleError);
   }
 
   public post(url: string, body: any, requestOptions?: RequestOptionsArgs): Observable<Response> {
-    return this.http.post(url, body, requestOptions ? requestOptions : HttpClient.options());
+    return this.http.post(url, body, requestOptions ? requestOptions : HttpClient.options())
+      .catch(this.handleError);
   }
 
   public put(url: string, body: any, requestOptions?: RequestOptionsArgs): Observable<Response> {
-    return this.http.put(url, body, requestOptions ? requestOptions : HttpClient.options());
+    return this.http.put(url, body, requestOptions ? requestOptions : HttpClient.options())
+      .catch(this.handleError);
   }
 
   public delete(url: string, requestOptions?: RequestOptionsArgs): Observable<Response> {
-    return this.http.delete(url, requestOptions);
+    return this.http.delete(url, requestOptions)
+      .catch(this.handleError);
   }
 
   public head(url: string, requestOptions?: RequestOptionsArgs): Observable<Response> {
-    return this.http.head(url, requestOptions);
+    return this.http.head(url, requestOptions)
+      .catch(this.handleError);
   }
 
   public patch(url: string, body: any, requestOptions?: RequestOptionsArgs): Observable<Response> {
-    return this.http.patch(url, body, requestOptions ? requestOptions : HttpClient.options());
+    return this.http.patch(url, body, requestOptions ? requestOptions : HttpClient.options())
+      .catch(this.handleError);
   }
 
   public request(url: string | Request, requestOptions?: RequestOptionsArgs): Observable<Response> {
-    return this.http.request(url, requestOptions);
+    return this.http.request(url, requestOptions)
+      .catch(this.handleError);
   }
 
   private static options() {
     let headers = new Headers({'Content-Type': 'application/json'});
     return new RequestOptions({headers: headers});
+  }
+
+  private handleError(error: Response | any) {
+    // In a real world app, you might use a remote logging infrastructure
+    let errMsg: string;
+    if (error instanceof Response) {
+      errMsg = `${error.statusText || 'Error!'}`;
+    } else {
+      errMsg = error.message ? error.message : error.toString();
+    }
+    console.error(errMsg);
+    return Observable.throw(errMsg);
   }
 
 }
