+++ src/main/scripts/src/app/settings/project/project.service.ts
@@ -25,8 +25,6 @@ import {Project} from '../../shared/dto/project';
 import {Observable} from 'rxjs/Observable';
 import {HttpClient} from '../../core/http-client.service';
 import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/observable/throw';
 
 @Injectable()
 export class ProjectService {
@@ -36,41 +34,24 @@ export class ProjectService {
 
   public getProjects(orgCode: string): Observable<Project[]> {
     return this.httpClient.get(ProjectService.apiPrefix(orgCode))
-      .map(response => response.json() as Project[])
-      .catch(this.handleError);
+      .map(response => response.json() as Project[]);
   }
 
   public getProject(orgCode: string, projCode: string): Observable<Project> {
     return this.httpClient.get(ProjectService.apiPrefix(orgCode, projCode))
-      .map(response => response.json() as Project)
-      .catch(this.handleError);
+      .map(response => response.json() as Project);
   }
 
   public deleteProject(orgCode: string, projCode: string): Observable<Response> {
-    return this.httpClient.delete(ProjectService.apiPrefix(orgCode, projCode))
-      .catch(this.handleError);
+    return this.httpClient.delete(ProjectService.apiPrefix(orgCode, projCode));
   }
 
   public createProject(orgCode: string, project: Project): Observable<Response> {
-    return this.httpClient.post(ProjectService.apiPrefix(orgCode), JSON.stringify(project))
-      .catch(this.handleError);
+    return this.httpClient.post(ProjectService.apiPrefix(orgCode), JSON.stringify(project));
   }
 
   public editProject(orgCode: string, projCode: string, project: Project): Observable<Response> {
-    return this.httpClient.put(ProjectService.apiPrefix(orgCode, projCode), JSON.stringify(project))
-      .catch(this.handleError);
-  }
-
-  private handleError(error: Response | any) {
-    // In a real world app, you might use a remote logging infrastructure
-    let errMsg: string;
-    if (error instanceof Response) {
-      errMsg = `${error.statusText || 'Error!'}`;
-    } else {
-      errMsg = error.message ? error.message : error.toString();
-    }
-    console.error(errMsg);
-    return Observable.throw(errMsg);
+    return this.httpClient.put(ProjectService.apiPrefix(orgCode, projCode), JSON.stringify(project));
   }
 
   private static apiPrefix(orgCode: string, projCode?: string): string {
